Move CryptoCore to it's own folder.
Crypto classes will be placed in core/crypto.
