+++ core/crypto/crypto_core.cpp
