+++ core/crypto/crypto_core.h
