+++ core/os/file_access.cpp
@@ -30,9 +30,9 @@
 
 #include "file_access.h"
 
+#include "core/crypto/crypto_core.h"
 #include "core/io/file_access_pack.h"
 #include "core/io/marshalls.h"
-#include "core/math/crypto_core.h"
 #include "core/os/os.h"
 #include "core/project_settings.h"
