+++ modules/websocket/wsl_peer.cpp
@@ -35,7 +35,7 @@
 #include "wsl_client.h"
 #include "wsl_server.h"
 
-#include "core/math/crypto_core.h"
+#include "core/crypto/crypto_core.h"
 #include "core/math/random_number_generator.h"
 #include "core/os/os.h"
