+++ platform/uwp/export/export.cpp
@@ -30,9 +30,9 @@
 
 #include "export.h"
 #include "core/bind/core_bind.h"
+#include "core/crypto/crypto_core.h"
 #include "core/io/marshalls.h"
 #include "core/io/zip_io.h"
-#include "core/math/crypto_core.h"
 #include "core/object.h"
 #include "core/os/dir_access.h"
 #include "core/os/file_access.h"
